blenderkit conversion moved to task
